+++ src/forms/products/styles.scss
@@ -1,3 +1,9 @@
+.button-form{
+  border-radius: 30px;
+  background-color: red;
+  border:solid;
+}
+
 .h1 {
   margin-bottom: 2.5rem;
   font-size: xx-large;
