+++ src/pages/ProductCreate.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import FormProduct from '../forms/products/form-product'
+
+const ProductCreate = () => {
+  return (
+    <>
+        <FormProduct />
+    </>
+  )
+}
+
+export default ProductCreate;
