+++ src/pages/Products.js
@@ -1,10 +1,15 @@
 import FormProduct from "../forms/products";
+import {useNavigate} from 'react-router-dom'
+import { Add } from '../forms/products';
+import ProductCreate from "./ProductCreate";
 
 
 
 const Products = () => {
+    const navigate = useNavigate()
+    
     return(
-            <FormProduct />
+            <Add onClick={()=> navigate('/products/create')} />
     )
 }
